requirements.txt
# Google APIs
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
google-generativeai>=0.3.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.1.0
numpy>=1.24.0
pytesseract>=0.3.10

# Text Processing
symspellpy>=6.7.0
pyspellchecker>=0.7.2
vncorenlp>=1.0.3

# Utils
python-dotenv>=1.0.0
schedule>=1.2.0
tenacity>=8.2.0
pandas>=2.1.0

# Additional Dependencies
requests>=2.31.0
tqdm>=4.66.0
regex>=2023.0.0